+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Swapped default values for `DropDownStyle.padding` and `DropDownStyle.listPadding`.
 * Added basic test suite.
 * Added `DropDownStyle.emptyListWidget` and `DropDownStyle.emptyListText` to display if the dropdown does not have any items.
+* Added `DropDownStyle.emptySearchResultsWidgetBuilder` and `DropDownStyle.emptySearchResultsTextBuilder` to display if the dropdown has no search results.
 
 ## 0.0.6